[+]add constraints
